requirements.txt
# Web Scraping
selenium==4.27.1
webdriver-manager==4.0.2
requests>=2.31.0
beautifulsoup4==4.12.3
lxml==5.3.0

# Database
psycopg2-binary==2.9.10
sqlalchemy==2.0.36

# Utilities
python-dotenv==1.0.1

# Compression handling
brotli>=1.1.0

# Optional: For better HTML parsing
html5lib==1.1
